requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
streamlit==1.35.0
langchain==0.2.6
langchain-openai==0.1.14
langchain-community==0.2.6
langchain-groq>=0.1.9
pandas==2.2.2
matplotlib==3.9.0
plotly==5.22.0
python-dotenv==1.0.1
pydantic==2.7.3
requests==2.32.3
pillow==10.3.0
